client/requirements.txt
gql[requests]
pyyaml
